dsa: Do not process packets with TCP/UDP if its iface is lan[123]
The Zephyr eth stack is not (by default) prepared to work with RAW and
TCP/UDP sockets enabled in the same time.

In this case the decision about forwarding is based on the destination
MAC address. If we find MAC, which we are looking for, then the TPC/UDP
stack is not used.

Signed-off-by: Lukasz Majewski <lukma@denx.de>