eth: t1s: Use temporary variable to properly store read value
The oa_tc6_reg_read() function accepts the pointer to 32 bit variable to
store read data. However, the oa_tc6_mdio_read_c45() function provides
pointer to data variable with only 16 bit length.

The problem is when the latter data is explicitly cast to uint32_t pointer
when calling oa_tc6_reg_read(). As a result the read value is stored
to 32 bit location overwriting the upper 2 bytes of memory.

As a result the stack is corrupted and program hangs.

To fix this problem - a temporary uint32_t variable is used to read the
value and then it is cast to uint16_t and stored to the data pointer.

Signed-off-by: Lukasz Majewski <lukma@denx.de>